pyproject.toml
[tool.poetry]
name = "ATRI"
version = "0.1.0"
description = "ATRI"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_remake", "nonebot_plugin_gachalogs", "nonebot_plugin_gspanel", "nonebot_plugin_tts_gal", "nonebot_plugin_wordcloud", "nonebot_plugin_tarot", "nonebot_plugin_crazy_thursday", "nonebot_plugin_covid19_news", "nonebot_plugin_apscheduler", "nonebot_plugin_fortune", "nonebot_plugin_logo", "leetcode", "nonebot_plugin_hitokoto", "nonebot_plugin_answersbook"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
